+++ DragonFruit.OnionFruit.Core.MacOS/Native/NetworkServiceInfo.cs
@@ -3,7 +3,7 @@
 
 using System.Runtime.InteropServices;
 
-namespace DragonFruit.OnionFruit.Core.MacOS.NativeStructs;
+namespace DragonFruit.OnionFruit.Core.MacOS.Native;
 
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 internal struct NetworkServiceInfo
